requirements.txt
fastapi
uvicorn
pydantic
slowapi
redis
motor
python-jose
bson
llama-parse
python-multipart
python-dotenv
httpx
pydantic[email]
passlib
pydantic-settings
passlib[bcrypt]
weasyprint
jinja2
google-cloud-storage
google-auth-oauthlib
google-auth-httplib2
google-api-python-client